pyproject.toml
[tool.poetry]
name = "qabot"
version = "0.1.0"
description = ""
authors = ["Ajit Singh <ajit.singh@uk.clara.net>"]

[tool.poetry.dependencies]
python = "^3.8.10"
Django = "^4.2"
djangorestframework = "^3.14.0"
langchain = "^0.0.133"
openai = "^0.27.4"
chromadb = "^0.3.21"
python-dotenv = "^1.0.0"
boto3 = "^1.26.118"
django-storages = {extras = ["boto3"], version = "^1.13.2"}
unstructured = "^0.6.2"
"pdfminer.six" = "^20221105"
python-jose = "^3.3.0"
djangorestframework-simplejwt = "^5.2.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^1.2.0"
black = "^23.3.0"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
    /(
        \.eggs/|
        \.git/|
        \.hg/|
        \.mypy_cache/|
        \.nox/|
        \.tox/|
        \.venv/|
        _build/|
        build/|
        dist/
    )/
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
